--- tests/simulator.test.ts
@@ -1,135 +1,0 @@
-import type {
-  CreatePromiseReq,
-  CreatePromiseRes,
-  DurablePromiseRecord,
-  RequestMsg,
-  ResponseMsg,
-} from "../src/network/network";
-import { Server } from "../src/server";
-import * as util from "../src/util";
-import { type Address, Message, Process, Simulator, anycast, unicast } from "../src/simulator";
-
-class Worker extends Process {
-  tick(time: number, messages: Message<ResponseMsg>[]): Message<RequestMsg>[] {
-    this.log(`tick ${time}`);
-    for (const msg of messages) {
-      this.log(msg);
-    }
-    // randomly sends a Request.
-    const req: CreatePromiseReq = {
-      kind: "createPromise",
-      id: "foo",
-      timeout: Number.MAX_SAFE_INTEGER,
-      tags: { "resonate:invoke": "local://any@group" },
-      iKey: "foo",
-    }; // if we remove the Ikey a weird assertion happens
-    return new Array<Message<CreatePromiseReq>>(new Message(unicast("server"), req, { requ: true }));
-  }
-}
-
-class Srvr extends Process {
-  private server: Server;
-
-  constructor(
-    public readonly iaddr: string,
-    public readonly gaddr: string[] = [],
-  ) {
-    super(iaddr, gaddr);
-    this.server = new Server();
-  }
-  tick(
-    time: number,
-    messages: Message<RequestMsg>[],
-  ): Message<
-    | ResponseMsg
-    | {
-        recv: string;
-        msg:
-          | { kind: "invoke" | "resume"; id: string; counter: number }
-          | { kind: "notify"; promise: DurablePromiseRecord };
-      }
-  >[] {
-    const resps: Message<
-      | ResponseMsg
-      | {
-          recv: string;
-          msg:
-            | { kind: "invoke" | "resume"; id: string; counter: number }
-            | { kind: "notify"; promise: DurablePromiseRecord };
-        }
-    >[] = new Array();
-
-    this.log(`tick ${time}`);
-
-    for (const msg of messages) {
-      switch (msg.data.kind) {
-        case "createPromise":
-          resps.push(msg.resp(anycast("default"), this.createPromise(msg.data)));
-          break;
-        default:
-          throw new Error(`Unsupported request kind: ${(msg.data as any).kind}`);
-      }
-      // append all responses and then return
-    }
-    // get msg to send.
-    const step = this.server.step(time);
-    let next = true;
-    let result: IteratorResult<{
-      recv: string;
-      msg:
-        | { kind: "invoke" | "resume"; id: string; counter: number }
-        | { kind: "notify"; promise: DurablePromiseRecord };
-    }>;
-
-    do {
-      result = step.next(next);
-      if (!result.done) {
-        const url = new URL(result.value.recv);
-        util.assert(url.protocol === "local:");
-        let target: Address;
-        if (url.username === "any") {
-          target =
-            url.pathname === "" ? anycast(url.hostname) : anycast(url.hostname, `${url.hostname}${url.pathname}`);
-        } else if (url.username === "uni") {
-          target = unicast(`${url.hostname}${url.pathname}`);
-        } else {
-          throw new Error(`not handled ${url}`);
-        }
-
-        resps.push(new Message(target, result.value));
-        next = true; // true or false
-      }
-    } while (!result.done);
-
-    return resps;
-  }
-
-  private createPromise(request: CreatePromiseReq): CreatePromiseRes {
-    return {
-      kind: "createPromise",
-      promise: this.server.createPromise(
-        request.id,
-        request.timeout,
-        request.param,
-        request.tags,
-        request.iKey,
-        request.strict,
-      ),
-    };
-  }
-}
-
-const s = new Simulator(0);
-
-// Server
-s.register(new Srvr("server"));
-
-// Workers
-s.register(new Worker("group/1", ["group"]));
-s.register(new Worker("group/0", ["group"]));
-
-let i = 0;
-while (i <= 2) {
-  s.tick();
-  i++;
-}
